% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bowron_attribs}
\alias{Bowron_attribs}
\title{Attributes for the Bowron CU}
\format{
A named vector with elements:
\describe{
  \item{CU_ID}{ID}
  \item{CU_Name}{Name}
  \item{AvGen}{Average Generation Length}
  \item{DataQualkIdx}{Describes the type of abundance data}
  \item{Cyclic}{Cycling dominance}
  \item{Cyc_Dom_Year}{If the CU has cyclic dominance, first dominant cycle year in the abundance data}
  \item{RelAbd_LBM}{Lower benchmark for the RelAbd metric}
  \item{RelAbd_UBM}{Upper benchmark for the RelAbd metric}
  \item{AbsAbd_LBM}{Lower benchmark for the AbsAbd metric}
  \item{AbsAbd_UBM}{Upper benchmark for the AbsAbd metric}
  \item{LongTrend_LBM}{Lower benchmark for the LongTrend metric}
  \item{LongTrend_UBM}{Upper benchmark for the LongTrend metric}
  \item{PercChange_LBM}{Lower benchmark for the PercChange metric}
  \item{PercChange_UBM}{Upper benchmark for the PercChange metric}
}
}
\usage{
Bowron_attribs
}
\description{
Attributes for the Bowron CU
}
\keyword{datasets}
