% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bowron_escapement}
\alias{Bowron_escapement}
\title{Escapement data for the Bowron CU}
\format{
A data frame with columns:
\describe{
  \item{CU_ID}{ID}
  \item{Year}{Data year}
  \item{Escapement_Wild}{Wild spawning escapment}
  }
}
\usage{
Bowron_escapement
}
\description{
Escapement data for the Bowron CU
}
\keyword{datasets}
