% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bowron_metrics}
\alias{Bowron_metrics}
\title{Metrics for the Bowron CU}
\format{
A data frame with columns:
\describe{
  \item{CU_ID}{ID}
  \item{Year}{Data year}
  \item{RelUBM}{RelUBM metric}
  \item{RelUBM.Status}{CU status as measured by the RelUBM metric}
  \item{RelUBM.Confidence}{Confidence in the RelUBM status}
  \item{AbsUBM}{AbsUBM metric}
  \item{AbsUBM.Status}{CU status as measured by the AbsUBM metric}
  \item{AbsUBM.Confidence}{Confidence in the AbsUBM status}
  \item{RelLBM}{RelLBM metric}
  \item{RelLBM.Status}{CU status as measured by the RelLBM metric}
  \item{RelLBM.Confidence}{Confidence in the RelLBM status}
  \item{AbsLBM}{AbsLBM metric}
  \item{AbsLBM.Status}{CU status as measured by the AbsLBM metric}
  \item{AbsLBM.Confidence}{Confidence in the AbsLBM status}
  \item{LongTrend}{LongTrend metric}
  \item{LongTrend.Status}{CU status as measured by the LongTrend metric}
  \item{LongTrend.Confidence}{Confidence in the LongTrend status}
  \item{PercChange}{PercChange metric}
  \item{PercChange.Status}{CU status as measured by the PercChange metric}
  \item{PercChange.Confidence}{Confidence in the PercChange status}
  \item{IntStatus.Status}{Status determined by Integrated Status Assessment}
  \item{IntStatus.Confidence}{Confidence in the Integrated Status}
  \item{IntStatus.Year}{Year the Integrated Status was last determined}
  \item{IntStatusRaw.Status}{Integrated Status, only for assessment years}
  \item{RapidStatus}{not used, included to allow for automatic processing of metrics file}
  \item{RapidStatus.Status}{Status determined by Rapid Status Assessment}
  \item{RapidStatus.Confidence}{Confidence in the Rapid Status}
}
}
\usage{
Bowron_metrics
}
\description{
Metrics for the Bowron CU
}
\keyword{datasets}
