% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{do.call2}
\alias{do.call2}
\title{An extended version of do.call}
\usage{
do.call2(fun, args, override)
}
\arguments{
\item{fun}{the function to call}

\item{args}{a named list that specifies the arguments for the function by name}

\item{override}{a named list with additional arguments; if arguments from the args list are included here,
the version in override will be used instead}
}
\value{
the response of the call to fun with the arguments specified
}
\description{
This version allows the user to specify a second set of args. If provided, the second set of args is added to the first set
and will override elements in the first set of args with the same name. Useful for letting the user pass on additional arguments to library
functions with lots of optional arguments.
}
\examples{
custom_plot <- function(x = 1:10, y = 1:10, additionalArgs = NULL) {
  do.call2(plot, args=list(x=x, y=y, xlab='x label', ylab='y label'), override=additionalArgs)
}
# show the default version of the plot
custom_plot()
# show a customized version
plotArgs <- list(main = 'A Custom Plot', xlab = 'new x label', ylab = 'new y label')
custom_plot(additionalArgs = plotArgs)

}
\seealso{
\code{\link[base]{do.call}}
}
