% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{do.plot}
\alias{do.plot}
\title{Run a graphics command with args and additional opts that may override args}
\usage{
do.plot(fun, args, override)
}
\arguments{
\item{fun}{the function to call}

\item{args}{a named list that specifies the arguments for the function by name}

\item{override}{a named list with additional arguments; if arguments from the args list are included here,
the version in override will be used instead}
}
\value{
the response of the call to fun with the arguments specified
}
\description{
Run a graphics command with args and additional opts that may override args
}
