% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fraser_sockeye_groups}
\alias{Fraser_sockeye_groups}
\title{Groups for Fraser sockeye CUs}
\format{
A list of lists that specifies group membership for Fraser sockeye CUs
}
\usage{
Fraser_sockeye_groups
}
\description{
Groups for Fraser sockeye CUs
}
\keyword{datasets}
