% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CU_summary_panel.R
\name{LogRelAbd_plot}
\alias{LogRelAbd_plot}
\title{Generate a Relative Abundance Plot (log scale)}
\usage{
LogRelAbd_plot(data, attribs, yrRange = NULL, gpar = metricPlotSpecs.default)
}
\arguments{
\item{data}{the Absolute Abundance time series (a named vector, names should be years)}

\item{attribs}{a list with CU attributes
if AvGen is provided, a running generational average will be added
if Cyclic is set and DomCycleYear is provided, values for dominant cycle years will be highlighted}

\item{yrRange}{range of years to show on x axis, given as a list with optional entries ming and/or max, i.e.
yrRange = list(min=<start year>, max=<end year>)}

\item{gpar}{plot styling}
}
\value{
generates a plot on the current graphics device
}
\description{
Generate a Relative Abundance Plot (log scale)
}
