% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CU_summary_panel.R
\name{metric_plot}
\alias{metric_plot}
\title{Generate a metric plot
with an optional running generational mean, color bands showing red, amber and green zones,
and optional additional horizontal and vertical reference lines}
\usage{
metric_plot(
  data,
  bmZones = NULL,
  avGen = NULL,
  domCycleYear = NULL,
  hRefLines = NULL,
  xRange = NULL,
  yRange = NULL,
  xTickStep = NULL,
  useLog = FALSE,
  plotTitle = "",
  tsName = "",
  vRefLines = NULL,
  gpar = metricPlotSpecs.default
)
}
\arguments{
\item{data}{the metric time series (a named vector, names should be years)}

\item{bmZones}{a list with entries Red and Amber, specifying the maximum values for each color zone}

\item{avGen}{if provided, a running average spanning avGen years will be added to the plot}

\item{domCycleYear}{first dominant cycle year; if set, indicates that CU should be treated as cyclic,
which will result in values for dominant years being highlighted}

\item{hRefLines}{a list with entries of format label = y. If provided, additional reference lines
will be shown at each of the given y values, with associated labels appearing near the right margin}

\item{xRange}{an optional list with entries min and/or max;
specifies the range of years to show on x axis; if min and/or max are specified,
this will override min and max values taken from the data}

\item{yRange}{range of values to show on the y axis, given as a vector with one or more elements}

\item{xTickStep}{if specified, force the number of years between tickmarks on x axis}

\item{useLog}{if true, show data on log scale}

\item{plotTitle}{the plot title}

\item{tsName}{the label to use for the metric time series}

\item{vRefLines}{a number or vector. If provided, draw a vertical line at the corresponding years in the graph}

\item{gpar}{plot styling}
}
\value{
generates a plot on the current graphics device
}
\description{
Generate a metric plot
with an optional running generational mean, color bands showing red, amber and green zones,
and optional additional horizontal and vertical reference lines
}
