% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CU_summary_panel.R
\name{metric_summary_pane}
\alias{metric_summary_pane}
\title{CU Metrics Summary Pane}
\usage{
metric_summary_pane(
  CUmetrics,
  CUts,
  CUattribs,
  metricStartYr = NULL,
  metricEndYr = NULL,
  metricPlotSpecs = metricPlotSpecs.default(),
  timelineMetricSpecs = timelineMetrics.default(),
  timelinePlotSpecs = timelinePlotSpecs.default()
)
}
\arguments{
\item{CUmetrics}{a data frame with CU metrics in columns; at a minimum, should contain a Year column
plus the columns specified in timelineMetrics}

\item{CUts}{a data frame with spawner abundance time series; at a minimum, should contain columns
Year, SpnForAbd_Wild}

\item{CUattribs}{a list of attributes; at a minimum, should contain the following fields
CU_Name: the name of the CU
RelAbd_LBM, RelAbd_UBM, AbsAbd_LBM, AbsAbd_UBM, LongTrend_LBM, LongTrend_UBM, PercChange_LBM, PercChange_UBM: upper and lower benchmarks for the status metrics
AvGen: average generation length
domCycleYear: the first dominant cycle year if the CU is cyclic
DataQualkIdx: 'AbsAbd' if the SpnForAbd_Wild time series has absolute abundance values
              'Rel.Idx' if the SpnForAbd_Wild time series is an index time series}

\item{metricStartYr}{optional: first year to show metrics for in timeline plot; also marked with a vertical reference line in the metric plots.}

\item{metricEndYr}{optional: final year to show metrics for in the metric plots; use this for consistent styling. Has no effect on timeline plot.}

\item{metricPlotSpecs}{optional:
layout options for the metrics plots - see \link{metricPlotSpecs.default} for details}

\item{timelineMetricSpecs}{optional:
If NULL, no timeline metrics are plotted.
If provided, timelineMetricSpecs should be a list of lists,
where each entry specifies one metric to include in the timeline plot
entries are specified as lists, containing a label and dataCol,
where dataCol is the name of the corresponding column in the CUmetrics data frame
and label is the label to show on the plot
the list may also contain a font attribute, which, if specified, overrides the default
the predefined timelineMetrics.default() structure contains the necessary definitions for working
with the data frame provided by the Scanner's DataManager module via get_metricSeries_for_CU(CU_id)
See \link{timelineMetrics.default} for more details.}

\item{timelinePlotSpecs}{optional:
layout options for the timeline plots - see \link{timelinePlotSpecs.default} for details}
}
\value{
generates a plot on the current graphics device
}
\description{
Creates metrics summary pane for a single CU. By default,
plots time series abundance, long-term trend, and percent-change metrics,
plus a timeline of status values for the different metrics.
The timeline of status values is optional and will be left out of timelineMetricSpecs is set to NULL.
}
\examples{
metric_summary_pane(Bowron_metrics, Bowron_escapement, Bowron_attribs)
# again, but without timeline panel
metric_summary_pane(Bowron_metrics, Bowron_escapement, Bowron_attribs, timelineMetricSpecs = NULL)
}
\seealso{
\code{\link{metricPlotSpecs.default}}
\code{\link{timelineMetrics.default}}
\code{\link{timelinePlotSpecs.default}}
}
