% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CU_summary_panel.R
\name{metricPlotSpecs.default}
\alias{metricPlotSpecs.default}
\title{Default plot styling for use with metric_plot}
\usage{
metricPlotSpecs.default()
}
\description{
This data structure provides the default gpar settings for the \link{metric_plot} function.
An equivalent data structure with additional values can be provided to metric_plot
via the gpar parameter to further customize the plot. Values provided via gpar will
override the defaults set here.
\describe{
   \item{par}{Additional arguments to be passed to \link[graphics]{par}}
   \item{main}{Additional arguments to be passed to \link[base]{plot}}
   \item{x.axis}{Additional styling for the x-axis (passed to \link[graphics]{axis})}
   \item{y.axis}{Additional styling for the y-axis (passed to \link[graphics]{axis})}
   \item{hline}{Additional styling for horizontal reference lines (passed to \link[graphics]{abline})}
   \item{hline.text}{Additional text styling for labels for the horizontal reference lines (passed to \link[graphics]{text})}
   \item{vline}{Additional styling for vertical reference lines (passed to \link[graphics]{abline})}
   \item{metric}{Additional styling for the metric time series (passed to \link[graphics]{lines})}
   \item{av}{Additional styling for the moving average time series (passed to \link[graphics]{lines})}
   \item{dom}{Additional styling for the dominant cycle time series (passed to \link[graphics]{lines})}
   \item{title}{Additional styling for the plot title (passed to \link[graphics]{title})}
}
}
\examples{
#  Bowron spawning escapement
data('Bowron_escapement', package='SoSplots')
data('Bowron_attribs', package='SoSplots')
esc <- Bowron_escapement$Escapement_Wild
names(esc) <- as.character(Bowron_escapement$Year)
# Plot of Bowron spawning escapement, plain black line
metric_plot(esc, gpar = list(metric = list(col='black', type='l')))
# RelAbd plot for Bowron, change line styles and colors for metric and moving average
RelAbd_plot(esc, Bowron_attribs,
  gpar = list(metric = list(col='black', type='l'), av = list(col='lightgrey')))
}
\seealso{
\code{\link{metric_plot}}
\code{\link{RelAbd_plot}}
\code{\link{LogRelAbd_plot}}
\code{\link{LogAbsAbd_plot}}
\code{\link{LongTrend_plot}}
\code{\link{PercChange_plot}}
\code{\link{status_overview_plot}}
}
