% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RapidStatus}
\alias{RapidStatus}
\title{Rapid Status Summary for Fraser Sockeye}
\format{
A data frame with CUs (identified by their CU_ID) in columns and assessment years in rows
}
\usage{
RapidStatus
}
\description{
Rapid Status Summary for Fraser Sockeye
}
\keyword{datasets}
