% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sockeye_labels}
\alias{sockeye_labels}
\title{Labels for sockeye CUs}
\format{
A list that specifies labels for sockeye CU IDs
}
\usage{
sockeye_labels
}
\description{
Labels for sockeye CUs
}
\keyword{datasets}
