% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{status_color}
\alias{status_color}
\title{Generate a color value from a status or status confidence value}
\usage{
status_color(status, alpha = 1, withAlpha = T)
}
\arguments{
\item{status}{a character string that gives the status or status confidence;
valid status options are 'Green', 'AmberGreen', 'Amber', 'AmberRed', 'Red', 'DD'
valid status confidence options are 'High', 'Moderate', 'Low'
a default color is returned for unrecognized status values}

\item{alpha}{optional, and integer between 0 and 1, specifies the alpha channel (i.e., transparency) value}

\item{withAlpha}{boolean; if TRUE, color strings will include an alpha channel - note that this may not be supported on all devices}
}
\value{
a string representing the status color in hex format
}
\description{
Generate a color value from a status or status confidence value
}
