% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapid_Status_panel.R
\name{status_overview_plot}
\alias{status_overview_plot}
\title{Generate a stacked bar plot that shows the proportion of CUs with red vs amber vs green status over time}
\usage{
status_overview_plot(
  CUstatus,
  plotTitle = "All CUs in Rapid Status Scan",
  gpar = statusOverviewSpecs.default
)
}
\arguments{
\item{CUstatus}{a matrix or data frame with CUs in columns and years in rows
row names of CUstatus will be used to indicate years and be plotted along the x axis}

\item{plotTitle}{the plot title}

\item{gpar}{plot styling}
}
\value{
generates a plot on the current graphics device
}
\description{
Generate a stacked bar plot that shows the proportion of CUs with red vs amber vs green status over time
}
