% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapid_Status_panel.R
\name{statusOverviewSpecs.default}
\alias{statusOverviewSpecs.default}
\title{Default plot specs for use with status_overview_plot}
\usage{
statusOverviewSpecs.default()
}
\description{
This data structure provides the default gpar settings for the \link{status_overview_plot} function.
An equivalent data structure with additional values can be provided to status_overview_plot
via the gpar parameter to further customize the plot. Values provided via gpar will
override the defaults set here.
\describe{
 \item{par}{Additional arguments to be passed to \link[graphics]{par}}
 \item{main}{Additional arguments to be passed to \link[base]{plot}}
 \item{x.axis}{Additional styling for the x-axis (passed to \link[graphics]{axis})}
 \item{y.axis}{Additional styling for the y-axis (passed \link[graphics]{axis})}
 \item{legend}{Additional styling for the legend (passed to \link[graphics]{barplot} via the *args.legend* parameter)}
 \item{grid}{Additional styling for the grid lines (passed to \link[graphics]{abline})}
}
}
\examples{
statusOverviewSpecs.default()
}
\seealso{
\code{\link{status_overview_plot}}
}
