% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timeline_Summary_Panel.R
\name{timeline_summary_plot}
\alias{timeline_summary_plot}
\title{Plot a summary of status time series for a set of CUs.
Status series are represented by series of squares, with letters and colouring indicating status
Each row in the block corresponds to a timeline for an individual CU.
CUs are grouped into blocks, with horizontal spaces between blocks and group labels shown above each block.}
\usage{
timeline_summary_plot(
  data,
  groups,
  labels,
  title = "Status Summary",
  selectedCUs = NULL,
  startYr = NULL,
  endYr = NULL,
  gpar = NULL,
  gparSelected = NULL
)
}
\arguments{
\item{data}{a data frame containing the status time series, years in rows, CU IDs in columns}

\item{groups}{a list of groups, each of which should consist of the CU IDs of all the CUs belonging to the group.
The name of the group will be used to label the group in the plot.
list(group1 = c('CU_ID1', 'CU_ID2', ...), group2 = c('CU_ID5', 'CU_ID6', ...))}

\item{labels}{a named list with labels for the CU IDs}

\item{title}{the plot title (optional)}

\item{selectedCUs}{a list of IDs of the currently selected CUs (optional)}

\item{startYr}{optional: if provided, the first year to show in the timeline}

\item{endYr}{optional: if provided the final year to show in the timeline}

\item{gpar}{plot styling - see timelineSummaryPlotSpecs.default() for details}

\item{gparSelected}{styling for selected CUs - see timelineSummaryPlotSpecsSelected.default() for details}
}
\value{
generates a plot on the current graphics device
}
\description{
Plot a summary of status time series for a set of CUs.
Status series are represented by series of squares, with letters and colouring indicating status
Each row in the block corresponds to a timeline for an individual CU.
CUs are grouped into blocks, with horizontal spaces between blocks and group labels shown above each block.
}
\examples{
data('RapidStatus', package='SoSplots')
data('Fraser_sockeye_groups', package='SoSplots')
data('sockeye_labels', package='SoSplots')
timeline_summary_plot(data=RapidStatus, group=Fraser_sockeye_groups, labels=sockeye_labels, title='Fraser Sockeye' )
}
