% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CU_summary_panel.R
\name{timelineMetrics.default}
\alias{timelineMetrics.default}
\title{Metrics to plot in the timeline_plot}
\usage{
timelineMetrics.default()
}
\description{
This data structure is predefined here for use with the data frame returned by the Salmon Scanner's DataManager
i.e., for use with the data frame returned by dataMngr$get_metricSeries_for_CU(CU_id).
More generally, it should be a list with one entry per metric, where each entry should be a list containing
a *label* and a *dataCol*, where *dataCol* is the name of the corresponding metric in the data frame
passed as the *data* arg to timeline_plot
}
\examples{
timelineMetrics.default()
# only show Rapid and Integrated Status in timeline_plot, make status letters bold
metrics <- list(
  list(label = "RapidStatus", dataCol = "RapidStatus.Status", font = 'bold'),
  list(label = "ConfRating", dataCol = "RapidStatus.Confidence"),
  list(label = "IntStatus", dataCol = "IntStatusRaw.Status", font = 'bold')
)

timeline_plot(Bowron_metrics, metrics = metrics)
}
\seealso{
\code{\link{timeline_plot}}
\code{\link{status_overview_plot}}
}
