% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CU_summary_panel.R
\name{timelinePlotSpecs.default}
\alias{timelinePlotSpecs.default}
\title{Default plot styling for use with timeline_plot}
\usage{
timelinePlotSpecs.default()
}
\description{
This data structure provides the default gpar settings for the \link{timeline_plot} function.
An equivalent data structure with additional values can be provided to timeline_plot
via the gpar parameter to further customize the plot. Values provided via gpar will
override the defaults set here.
\describe{
 \item{par}{Additional arguments to be passed to \link[graphics]{par}}
 \item{main}{Additional arguments to be passed to \link[base]{plot}}
 \item{x.axis}{Additional styling for the x-axis (passed \link[graphics]{axis})}
 \item{label}{Additional styling for the metric series labels (passed to \link[graphics]{text})}
 \item{metric.text}{Additional text styling for the letters representing status values (passed to \link[graphics]{text})}
 \item{metric.line}{Additional styling for the horizontal lines shown as the backdrop for each metric series (passed to \link[graphics]{abline})}
}
}
\examples{
timelinePlotSpecs.default()
# show timeline plot with new title styled in red
timeline_plot(Bowron_metrics, gpar=list(title=list(text='Bowron Status', col='red', cex=1.2)))
}
\seealso{
\code{\link{timeline_plot}}
\code{\link{status_overview_plot}}
}
