% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timeline_Summary_Panel.R
\name{timelineSummaryPlotSpecs.default}
\alias{timelineSummaryPlotSpecs.default}
\title{Default plot styling for use with timeline_summary_plot}
\usage{
timelineSummaryPlotSpecs.default()
}
\description{
This data structure provides the default gpar settings for the \link{timeline_summary_plot} function.
An equivalent data structure with additional values can be provided to timeline_summary_plot
via the gpar parameter to further customize the plot. Values provided via gpar will
override the defaults set here.
\describe{
 \item{par}{Additional arguments to be passed to \link[graphics]{par}}
 \item{main}{Additional arguments to be passed to \link[base]{plot}}
 \item{x.axis}{Additional styling for the x-axis (passed \link[graphics]{axis})}
 \item{label}{Additional styling for the CU time series labels (passed to \link[graphics]{text})}
 \item{status.text}{Additional text styling for the letters representing status values (passed to \link[graphics]{text})}
 \item{status.line}{Additional styling for the horizontal lines shown as the backdrop for each status series (passed to \link[graphics]{abline})}
}
}
\examples{
timelineSummaryPlotSpecs.default()
# show timeline summary plot with new title styled in red
data('RapidStatus', package='SoSplots')
data('Fraser_sockeye_groups', package='SoSplots')
data('sockeye_labels', package='SoSplots')
timeline_summary_plot(data=RapidStatus, group=Fraser_sockeye_groups, labels=sockeye_labels, gpar=list(title=list(text='Fraser Sockeye', col='red', cex=1.2)))
}
\seealso{
\code{\link{timeline_summary_plot}}
}
