% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timeline_Summary_Panel.R
\name{timelineSummaryPlotSpecsSelected.default}
\alias{timelineSummaryPlotSpecsSelected.default}
\title{Default plot styling for selected CUs, for use with timeline_summary_plot()}
\usage{
timelineSummaryPlotSpecsSelected.default()
}
\description{
This data structure provides the default gparSelected settings for the \link{timeline_summary_plot} function.
An equivalent data structure can be provided to timeline_summary_plot via the gparSelect parameter to further customize the plot.
Values provided via the gpar parameter will override the defaults set here.
\describe{
 \item{label}{Styling for the CU time series labels (passed to \link[graphics]{text})}
 \item{status.text}{Styling for the letters representing status values (passed to \link[graphics]{text})}
 \item{status.line}{Styling for the horizontal lines shown as the backdrop for each status series (passed to \link[graphics]{abline})}
}
}
\seealso{
\code{\link{timeline_summary_plot}}
\code{\link{timelineSummaryPlotSpecs.default}}
}
