% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RapidStatus}
\alias{RapidStatus}
\title{Rapid Status Summary for Fraser Sockeye
#'
A data frame with CUs (identified by their CU_ID) in columns and assessment years in rows}
\format{
An object of class \code{data.frame} with 25 rows and 22 columns.
}
\usage{
RapidStatus
}
\description{
Rapid Status Summary for Fraser Sockeye
#'
A data frame with CUs (identified by their CU_ID) in columns and assessment years in rows
}
\keyword{datasets}
