% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CU_summary_panel.R
\name{RelAbd_plot}
\alias{RelAbd_plot}
\title{Plot of Relative Abundance}
\usage{
RelAbd_plot(data, attribs, yrRange = NULL, gpar = NULL)
}
\arguments{
\item{data}{the abundance time series (a named vector, names should be years)}

\item{attribs}{a list with CU attributes RelAbd_LBM and RelAbd_UBM, the lower and upper benchmark respectively
if AvGen is provided, a running generational average will be added
if DomCycleYear is provided, values for dominant cycle years will be highlighted}

\item{yrRange}{range of years to show on x axis, given as a list with optional entries ming and/or max, i.e.
yrRange = list(min=<start year>, max=<end year>)}

\item{gpar}{plot styling - see metricPlotSpecs.default() for details}
}
\value{
generates a plot on the current graphics device
}
\description{
Generate a plot of a time series interpreted as relative abundance,
with color bands showing red, amber and green zones for the Relative Abundance metric.
}
\examples{
ts <- Bowron_escapement$Escapement_Wild
names(ts) <- as.character(Bowron_escapement$Year)
RelAbd_plot(ts, Bowron_attribs)
}
\seealso{
\code{\link{LogRelAbd_plot}}
\code{\link{metricPlotSpecs.default}}
\code{\link{metric_plot}}
}
