% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{status_letter}
\alias{status_letter}
\title{Generate a letter from a status color for use as an abbreviated description of the status}
\usage{
status_letter(status)
}
\arguments{
\item{status}{a character string that gives the status or status confidence;
valid status options are 'Green', 'AmberGreen', 'Amber', 'AmberRed', 'Red', 'DD'
valid status confidence options are 'High', 'Moderate', 'Low'}
}
\value{
a string containing one or more letters representing the status or status confidence
}
\description{
Generate a letter from a status color for use as an abbreviated description of the status
}
