% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapid_Status_panel.R
\name{status_overview_plot}
\alias{status_overview_plot}
\title{Status Overview Pane}
\usage{
status_overview_plot(
  CUstatus,
  plotTitle = "All CUs in Rapid Status Scan",
  gpar = NULL
)
}
\arguments{
\item{CUstatus}{a matrix or data frame with CUs in columns and years in rows
row names of CUstatus will be used to indicate years and be plotted along the x axis}

\item{plotTitle}{the plot title}

\item{gpar}{plot styling - see \link{statusOverviewSpecs.default} for details}
}
\value{
generates a plot on the current graphics device
}
\description{
Generate a stacked bar plot that shows the proportion of CUs with red vs amber vs green status over time.
}
\examples{
# show status overview for Fraser CUs
data('RapidStatus', package='SoSplots')
status_overview_plot(RapidStatus)
# again, with modified title
gpar <- list(main=list(main='Rapid Status for Fraser Sockeye CUs', cex.main=0.8))
status_overview_plot(RapidStatus, gpar=gpar)
}
