% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CU_summary_panel.R
\name{timeline_plot}
\alias{timeline_plot}
\title{Plot a set of time series using a series of squares, with letters and colouring indicating status (or confidence in status)}
\usage{
timeline_plot(
  data,
  metrics = timelineMetrics.default(),
  title = "Metrics & Status",
  startYr = NULL,
  endYr = NULL,
  gpar = NULL
)
}
\arguments{
\item{data}{a data frame containing the metrics to plot. Metrics should be in columns.
data must also contain a 'Year' column, to be used as the x-axis values.}

\item{metrics}{a list of lists specifying the metrics to plot. At a minimum, each entry in metrics must specify
a label attribute and a dataCol attribute, where dataCol provides the column name of the metric in data.
If a font attribute is provided, it overrides the default font.
Font options are 'plain' (the default), 'bold', 'italic', 'bold italic'.
Note: the font set here is used to select one of the fonts set via par('font'),
following the convention that font #1 corresponds to plain text, font #2 to bold etc.
See par documentation for details. Font can also be specified more directly by giving the numeric index.
See \link{timelineMetrics.default} for more details on how to specify what metrics should be shown.}

\item{title}{the plot title (optional)}

\item{startYr}{optional: if provided, the first year to show in the timeline}

\item{endYr}{optional: if provided the final year to show in the timeline}

\item{gpar}{plot styling - see timelinePlotSpecs.default() for details}
}
\value{
generates a plot on the current graphics device
}
\description{
Plot a set of time series using a series of squares, with letters and colouring indicating status (or confidence in status)
}
\examples{
# show all metrics
timeline_plot(Bowron_metrics)
metrics <- list(
  list(label = "RapidStatus" , dataCol = "RapidStatus.Status", font = 'bold'),
  list(label = "ConfRating", dataCol = "RapidStatus.Confidence"),
  list(label = "IntStatus", dataCol = "IntStatusRaw.Status", font = 'bold')
)
# only show Rapid and Integrated Status
timeline_plot(Bowron_metrics, metrics = metrics)
}
\seealso{
\code{\link{timelineMetrics.default}}
\code{\link{timelinePlotSpecs.default}}
}
